Poprawki strony manuala -- Kamil Jarosławski (kamil%visual.pl). Obsługa próbek o rozszerzeniu ogg -- Tomasz Marciniak (ciniol%atos.wmid.amu.edu.pl) Oficjalne włączenie poprawki Michała Maternika (michal_m%ks.k.pl) pozwalającej korzystać z niemieckiego słownika. Poprawka działania na amd64 -- Arkadiusza Kołacz (akolacz%gmail.com).

+++ src/ydpcore.h
@@ -25,7 +25,7 @@
 #define YDP_CANTOPENDEF 2
 #define YDP_INVALIDFILE 3
 
-unsigned long *indexes;	/* tablica z indeksami definicji w pliku indeksowym */
+unsigned int *indexes;	/* tablica z indeksami definicji w pliku indeksowym */
 u_char **words;		/* tablica z list± s³ów */
 unsigned short wordcount;	/* ilo¶æ s³ów */
 FILE *fd, *fi;		/* strumienie */
